refactor(e2e): extract shared raw-actor session helper
Pulls the raw-JWT Supabase client sign-in logic out of direct-write-denial.test.ts into test/e2e/raw-actor.ts, so the upcoming read-scoping spec can reuse it instead of duplicating it. No behavior change.
